requirements.txt
Django==5.2.7
django-environ==0.11.2
Pillow==10.4.0
